Capitalazing hyperparameters
