+++ Python Trainer/builds/distance_reward_old_env/selfDriving_Data/ML-Agents/Timers/Main_timers.json
@@ -1,0 +1,1 @@
+{"count":1,"self":231.1198464,"total":459.09932299999997,"children":{"InitializeActuators":{"count":4,"self":0.00090699999999999993,"total":0.00090699999999999993,"children":null},"InitializeSensors":{"count":4,"self":0.003209,"total":0.003209,"children":null},"AgentSendState":{"count":21356,"self":106.14760319999999,"total":119.161452,"children":{"CollectObservations":{"count":17088,"self":0.089693,"total":0.089693,"children":null},"WriteActionMask":{"count":17088,"self":0.042984999999999995,"total":0.042984999999999995,"children":null},"RequestDecision":{"count":17088,"self":0.081142999999999993,"total":12.881167999999999,"children":{"AgentInfo.ToProto":{"count":17088,"self":0.13647299999999998,"total":12.800025,"children":{"GenerateSensorData":{"count":17088,"self":0.908678,"total":12.663552,"children":{"CameraSensor.GetCompressedObservation":{"count":17088,"self":11.7548736,"total":11.754874,"children":null}}}}}}}}},"DecideAction":{"count":21356,"self":106.34604159999999,"total":106.346043,"children":null},"AgentAct":{"count":21356,"self":2.25724,"total":2.466876,"children":{"AgentInfo.ToProto":{"count":448,"self":0.0047279999999999996,"total":0.209636,"children":{"GenerateSensorData":{"count":448,"self":0.022701,"total":0.20490799999999998,"children":{"CameraSensor.GetCompressedObservation":{"count":448,"self":0.18220699999999998,"total":0.18220699999999998,"children":null}}}}}}}},"gauges":{"CarBehavior.CumulativeReward":{"count":448,"max":1936.69678,"min":-53.49982,"runningAverage":266.0384,"value":319.711548,"weightedAverage":231.423828}},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1710156657","unity_version":"2022.3.13f1","command_line_arguments":"\/home\/tadeas\/staz\/self-driving\/.\/Python Trainer\/builds\/distance_reward_old_env\/selfDriving.x86_64 --mlagents-port 5005","communication_protocol_version":"1.5.0","com.unity.ml-agents_version":"3.0.0-exp.1","scene_name":"Main","end_time_seconds":"1710157116"}}
